Registry actions received from openraft are processed and applied to state machine
