+++ src.old/types/registry_state.rs
@@ -18,73 +18,6 @@ use tokio::sync::Mutex;
 use super::MachineMetricLabels;
 use super::{BlobEntry, ManifestEntry};
 
-#[derive(Default)]
-struct Store {
-    blobs: HashMap<Digest, Blob>,
-    manifests: HashMap<Digest, Manifest>,
-    tags: HashMap<RepositoryName, HashMap<String, Digest>>,
-}
-
-impl Store {
-    fn get_mut_blob(&mut self, digest: &Digest, timestamp: DateTime<Utc>) -> Option<&mut Blob> {
-        if let Some(mut blob) = self.blobs.get_mut(digest) {
-            blob.updated = timestamp;
-            return Some(blob);
-        }
-
-        None
-    }
-
-    fn get_or_insert_blob(&mut self, digest: Digest, timestamp: DateTime<Utc>) -> &mut Blob {
-        let mut blob = self.blobs.entry(digest).or_insert_with(|| Blob {
-            created: timestamp,
-            updated: timestamp,
-            content_type: None,
-            size: None,
-            dependencies: Some(vec![]),
-            locations: HashSet::new(),
-            repositories: HashSet::new(),
-        });
-
-        blob.updated = timestamp;
-
-        blob
-    }
-
-    fn get_mut_manifest(
-        &mut self,
-        digest: &Digest,
-        timestamp: DateTime<Utc>,
-    ) -> Option<&mut Manifest> {
-        if let Some(mut manifest) = self.manifests.get_mut(digest) {
-            manifest.updated = timestamp;
-            return Some(manifest);
-        }
-
-        None
-    }
-
-    fn get_or_insert_manifest(
-        &mut self,
-        digest: Digest,
-        timestamp: DateTime<Utc>,
-    ) -> &mut Manifest {
-        let mut manifest = self.manifests.entry(digest).or_insert_with(|| Manifest {
-            created: timestamp,
-            updated: timestamp,
-            content_type: None,
-            size: None,
-            dependencies: Some(vec![]),
-            locations: HashSet::new(),
-            repositories: HashSet::new(),
-        });
-
-        manifest.updated = timestamp;
-
-        manifest
-    }
-}
-
 pub struct RegistryState {
     state: Mutex<Store>,
     pub events: tokio::sync::broadcast::Sender<RaftEvent>,
@@ -220,76 +153,6 @@ impl RegistryState {
         matches!(self.webhook_send.send(event).await, Ok(_))
     }
 
-    pub async fn is_blob_available(&self, repository: &RepositoryName, hash: &Digest) -> bool {
-        let store = self.state.lock().await;
-        match store.blobs.get(hash) {
-            None => false,
-            Some(blob) => blob.repositories.contains(repository),
-        }
-    }
-
-    pub async fn get_blob_directly(&self, hash: &Digest) -> Option<Blob> {
-        let store = self.state.lock().await;
-        store.blobs.get(hash).cloned()
-    }
-
-    pub async fn get_blob(&self, repository: &RepositoryName, hash: &Digest) -> Option<Blob> {
-        let store = self.state.lock().await;
-        match store.blobs.get(hash) {
-            None => None,
-            Some(blob) => {
-                if blob.repositories.contains(repository) {
-                    return Some(blob.clone());
-                }
-                None
-            }
-        }
-    }
-
-    pub async fn get_manifest_directly(&self, hash: &Digest) -> Option<Manifest> {
-        let store = self.state.lock().await;
-        store.manifests.get(hash).cloned()
-    }
-    pub async fn get_manifest(
-        &self,
-        repository: &RepositoryName,
-        hash: &Digest,
-    ) -> Option<Manifest> {
-        let store = self.state.lock().await;
-        match store.manifests.get(hash) {
-            None => None,
-            Some(manifest) => {
-                if manifest.repositories.contains(repository) {
-                    return Some(manifest.clone());
-                }
-                None
-            }
-        }
-    }
-    pub async fn get_tag(&self, repository: &RepositoryName, tag: &str) -> Option<Digest> {
-        let store = self.state.lock().await;
-        match store.tags.get(repository) {
-            Some(repository) => repository.get(tag).cloned(),
-            None => None,
-        }
-    }
-
-    pub async fn get_tags(&self, repository: &RepositoryName) -> Option<Vec<String>> {
-        let store = self.state.lock().await;
-        store
-            .tags
-            .get(repository)
-            .map(|repository| repository.keys().cloned().collect())
-    }
-
-    pub async fn is_manifest_available(&self, repository: &RepositoryName, hash: &Digest) -> bool {
-        let store = self.state.lock().await;
-        match store.manifests.get(hash) {
-            None => false,
-            Some(manifest) => manifest.repositories.contains(repository),
-        }
-    }
-
     pub async fn get_orphaned_blobs(&self) -> Vec<BlobEntry> {
         let store = self.state.lock().await;
 
@@ -354,188 +217,6 @@ impl RegistryState {
                 digest,
             })
             .collect::<Vec<ManifestEntry>>()
-    }
-
-    pub async fn dispatch_entries(&self, event: RaftEvent) {
-        match &event {
-            RaftEvent::Committed {
-                start_index,
-                entries,
-            } => {
-                self.handle_raft_commit(entries.clone()).await;
-
-                if let Some(start_index) = start_index {
-                    let labels = MachineMetricLabels {
-                        identifier: self.machine_identifier.clone(),
-                    };
-                    self.last_applied_index
-                        .get_or_create(&labels)
-                        .set(*start_index as u64);
-                }
-            }
-        }
-
-        // After we have handled the event we can pass it on to downstream
-        // subscribers. They can then be happy that the store is as up to date as needed
-        if let Err(err) = self.events.send(event) {
-            warn!("Error while notifying of commit events: {err:?}");
-        }
-    }
-
-    async fn handle_raft_commit(&self, actions: Vec<LogEntry>) {
-        let mut store = self.state.lock().await;
-
-        for action in actions {
-            match action.entry {
-                RegistryAction::Empty {} => {}
-                RegistryAction::BlobStored {
-                    timestamp,
-                    user: _,
-                    digest,
-                    location,
-                } => {
-                    let blob = store.get_or_insert_blob(digest.clone(), timestamp);
-                    blob.locations.insert(location.clone());
-
-                    if location == self.machine_identifier {
-                        self.blob_available(&digest).await;
-                    }
-                }
-                RegistryAction::BlobUnstored {
-                    timestamp,
-                    user: _,
-                    digest,
-                    location,
-                } => {
-                    if let Some(blob) = store.get_mut_blob(&digest, timestamp) {
-                        blob.locations.remove(&location);
-
-                        if blob.locations.is_empty() {
-                            store.blobs.remove(&digest);
-                        }
-                    }
-                }
-                RegistryAction::BlobMounted {
-                    timestamp,
-                    user: _,
-                    digest,
-                    repository,
-                } => {
-                    let blob = store.get_or_insert_blob(digest, timestamp);
-                    blob.repositories.insert(repository);
-                }
-                RegistryAction::BlobUnmounted {
-                    timestamp,
-                    user: _,
-                    digest,
-                    repository,
-                } => {
-                    if let Some(blob) = store.get_mut_blob(&digest, timestamp) {
-                        blob.repositories.remove(&repository);
-                    }
-                }
-                RegistryAction::BlobInfo {
-                    timestamp,
-                    digest,
-                    dependencies,
-                    content_type,
-                } => {
-                    if let Some(mut blob) = store.get_mut_blob(&digest, timestamp) {
-                        blob.dependencies = Some(dependencies);
-                        blob.content_type = Some(content_type);
-                    }
-                }
-                RegistryAction::BlobStat {
-                    timestamp,
-                    digest,
-                    size,
-                } => {
-                    if let Some(mut blob) = store.get_mut_blob(&digest, timestamp) {
-                        blob.size = Some(size);
-                    }
-                }
-                RegistryAction::ManifestStored {
-                    timestamp,
-                    user: _,
-                    digest,
-                    location,
-                } => {
-                    let manifest = store.get_or_insert_manifest(digest.clone(), timestamp);
-                    manifest.locations.insert(location.clone());
-
-                    if location == self.machine_identifier {
-                        self.manifest_available(&digest).await;
-                    }
-                }
-                RegistryAction::ManifestUnstored {
-                    timestamp,
-                    user: _,
-                    digest,
-                    location,
-                } => {
-                    if let Some(manifest) = store.get_mut_manifest(&digest, timestamp) {
-                        manifest.locations.remove(&location);
-
-                        if manifest.locations.is_empty() {
-                            store.manifests.remove(&digest);
-                        }
-                    }
-                }
-                RegistryAction::ManifestMounted {
-                    timestamp,
-                    user: _,
-                    digest,
-                    repository,
-                } => {
-                    let manifest = store.get_or_insert_manifest(digest, timestamp);
-                    manifest.repositories.insert(repository);
-                }
-                RegistryAction::ManifestUnmounted {
-                    timestamp,
-                    user: _,
-                    digest,
-                    repository,
-                } => {
-                    if let Some(manifest) = store.get_mut_manifest(&digest, timestamp) {
-                        manifest.repositories.remove(&repository);
-
-                        if let Some(tags) = store.tags.get_mut(&repository) {
-                            tags.retain(|_, value| value != &digest);
-                        }
-                    }
-                }
-                RegistryAction::ManifestInfo {
-                    timestamp,
-                    digest,
-                    dependencies,
-                    content_type,
-                } => {
-                    if let Some(mut manifest) = store.get_mut_manifest(&digest, timestamp) {
-                        manifest.dependencies = Some(dependencies);
-                        manifest.content_type = Some(content_type);
-                    }
-                }
-                RegistryAction::ManifestStat {
-                    timestamp,
-                    digest,
-                    size,
-                } => {
-                    if let Some(mut manifest) = store.get_mut_manifest(&digest, timestamp) {
-                        manifest.size = Some(size);
-                    }
-                }
-                RegistryAction::HashTagged {
-                    timestamp: _,
-                    user: _,
-                    digest,
-                    repository,
-                    tag,
-                } => {
-                    let repository = store.tags.entry(repository).or_insert_with(HashMap::new);
-                    repository.insert(tag, digest);
-                }
-            }
-        }
     }
 }
