+++ src/types/mod.rs
@@ -1,0 +1,11 @@
+mod actions;
+mod blob;
+mod digest;
+mod manifest;
+mod repository_name;
+
+pub(crate) use actions::RegistryAction;
+pub(crate) use blob::Blob;
+pub(crate) use digest::Digest;
+pub(crate) use manifest::Manifest;
+pub(crate) use repository_name::RepositoryName;
